move ControlProportionSettings from utils to models
