+++ app/models/BannerTests.scala
@@ -4,7 +4,6 @@ import enumeratum.{CirceEnum, Enum, EnumEntry}
 import io.circe.generic.extras.Configuration
 import io.circe.generic.extras.auto._
 import io.circe.{Decoder, Encoder}
-import utils.ControlProportionSettings
 
 import scala.collection.immutable.IndexedSeq
