+++ app/models/ControlProportionSettings.scala
@@ -1,3 +1,3 @@
-package utils
+package models
 
 case class ControlProportionSettings(proportion: Float, offset: Int)
